refactor: add accordion component
