+++ plugin/web/extensions/sidebar/core/components/content/common/DatasetCard/Field/index.tsx
@@ -1,18 +1,9 @@
 import { Group, Template } from "@web/extensions/sidebar/core/types";
 import { ReearthApi } from "@web/extensions/sidebar/types";
 import { postMsg } from "@web/extensions/sidebar/utils";
-import { Icon } from "@web/sharedComponents";
-import { styled } from "@web/theme";
 import { useCallback, useEffect, useMemo, useState } from "react";
-import {
-  Accordion,
-  AccordionItem,
-  AccordionItemHeading,
-  AccordionItemButton,
-  AccordionItemPanel,
-  AccordionItemState,
-} from "react-accessible-accordion";
 
+import AccordionComponent from "./AccordionComponent";
 import fields from "./Fields";
 import {
   ConfigData,
@@ -78,7 +69,6 @@ const FieldComponent: React.FC<Props> = ({
 }) => {
   const Field = fields[field.type];
   const [groupPopupOpen, setGroupPopup] = useState(false);
-  const [isExpanded, toggleExpanded] = useState(true);
 
   const handleGroupSelectOpen = useCallback(
     (e: React.MouseEvent<HTMLDivElement, MouseEvent> | undefined) => {
@@ -155,125 +145,34 @@ const FieldComponent: React.FC<Props> = ({
       onCurrentGroupUpdate={onCurrentGroupUpdate}
     />
   ) : (
-    <StyledAccordionComponent
-      allowZeroExpanded
-      preExpanded={[field.id]}
-      hide={!editMode && !Field?.hasUI}>
-      <AccordionItem uuid={field.id} dangerouslySetExpanded={isExpanded}>
-        <AccordionItemState>
-          {({ expanded }) => (
-            <Header showBorder={expanded}>
-              {editMode ? (
-                <HeaderContents>
-                  <LeftContents onClick={() => toggleExpanded(v => !v)}>
-                    {Field && (
-                      <ArrowIcon icon="arrowDown" size={16} direction="right" expanded={expanded} />
-                    )}
-                    <Title>{editModeTitle}</Title>
-                  </LeftContents>
-                  <RightContents>
-                    <StyledIcon icon="arrowUpThin" size={16} onClick={handleUpClick} />
-                    <StyledIcon icon="arrowDownThin" size={16} onClick={handleDownClick} />
-                    <StyledIcon
-                      icon="group"
-                      color={field.group ? "#00BEBE" : "inherit"}
-                      size={16}
-                      onClick={handleGroupSelectOpen}
-                    />
-                    <StyledIcon icon="trash" size={16} onClick={handleRemove} />
-                  </RightContents>
-                </HeaderContents>
-              ) : (
-                <HeaderContents>
-                  <Title>{title}</Title>
-                  <ArrowIcon icon="arrowDown" size={16} direction="left" expanded={expanded} />
-                </HeaderContents>
-              )}
-            </Header>
-          )}
-        </AccordionItemState>
-        {Field?.Component && (
-          <BodyWrapper>
-            <Field.Component
-              value={{ ...field }}
-              editMode={editMode}
-              isActive={isActive}
-              templates={templates}
-              fieldGroups={selectGroups}
-              configData={configData}
-              dataID={dataID}
-              onUpdate={onUpdate?.(field.id)}
-              onCurrentGroupUpdate={onCurrentGroupUpdate}
-              onSceneUpdate={onSceneUpdate}
-            />
-          </BodyWrapper>
-        )}
-      </AccordionItem>
-    </StyledAccordionComponent>
+    <AccordionComponent
+      id={field.id}
+      isGroup={!!field.group}
+      editMode={editMode}
+      hasUI={Field?.hasUI}
+      showArrowIcon={!!Field}
+      title={title}
+      editModeTitle={editModeTitle}
+      onGroupSelectOpen={handleGroupSelectOpen}
+      onRemove={handleRemove}
+      onUpClick={handleUpClick}
+      onDownClick={handleDownClick}>
+      {Field?.Component && (
+        <Field.Component
+          value={{ ...field }}
+          editMode={editMode}
+          isActive={isActive}
+          templates={templates}
+          fieldGroups={selectGroups}
+          configData={configData}
+          dataID={dataID}
+          onUpdate={onUpdate?.(field.id)}
+          onCurrentGroupUpdate={onCurrentGroupUpdate}
+          onSceneUpdate={onSceneUpdate}
+        />
+      )}
+    </AccordionComponent>
   );
 };
 
 export default FieldComponent;
-
-const StyledAccordionComponent = styled(Accordion)<{ hide: boolean }>`
-  ${({ hide }) => hide && "display: none;"}
-  width: 100%;
-  border: 1px solid #e6e6e6;
-  border-radius: 4px;
-  background: #ffffff;
-`;
-
-const Header = styled(AccordionItemHeading)<{ showBorder?: boolean }>`
-  border-bottom-width: 1px;
-  border-bottom-style: solid;
-  border-bottom-color: transparent;
-  ${({ showBorder }) => showBorder && "border-bottom-color: #e0e0e0;"}
-  display: flex;
-  height: auto;
-`;
-
-const HeaderContents = styled(AccordionItemButton)`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex: 1;
-  padding: 12px;
-  outline: none;
-  cursor: pointer;
-`;
-
-const BodyWrapper = styled(AccordionItemPanel)`
-  position: relative;
-  border-radius: 0px 0px 4px 4px;
-  padding: 12px;
-`;
-
-const Title = styled.p`
-  margin: 0;
-  user-select: none;
-  width: 160px;
-  overflow-wrap: break-word;
-`;
-
-const StyledIcon = styled(Icon)`
-  cursor: pointer;
-`;
-
-const LeftContents = styled.div`
-  display: flex;
-  align-items: center;
-`;
-
-const RightContents = styled.div`
-  display: flex;
-  gap: 4px;
-`;
-
-const ArrowIcon = styled(Icon)<{ direction: "left" | "right"; expanded?: boolean }>`
-  transition: transform 0.15s ease;
-  ${({ direction, expanded }) =>
-    (direction === "right" && !expanded && "transform: rotate(-90deg);") ||
-    (direction === "left" && !expanded && "transform: rotate(90deg);") ||
-    null}
-  ${({ direction }) => (direction === "left" ? "margin: 0 -4px 0 4px;" : "margin: 0 4px 0 -4px;")}
-`;
